Always show reset menu option in tablet
There is a bug in android that is preventing the reset from being shown whenset as hidden in xml.
This prevented the reset button from being shown after a sub region query, giveing the user no
means to reset the query.

Lets just show it all the time, resetting when already reset is a no-op